General improvement
Added JavaDoc to the class header.
Put in a check regarding the size of the hash table.
Added a missing initialization block for the hash table's lists.
Changed the add function to return a reference to the HashTable, much
like StringList does.
Relegated the "unique" handling of addUnique to the StringList.
Removed a commented-out function that was inherently flawed.