Improvements: TryLock(String), ShuffleArray; Removed enumerateArray
Removed the extension of BaseHandler now that the class has been
removed.
TryLock(String): added a signature for when we implement holdTime via
StringToStringDictionary.
EnumerateArray: removed the function now that
StringHandler.concatStringArray performs the same function.
ShuffleArray: fixed an issue where the "randomized" segment doesn't grow
with the target index.