Modify Function: Length(StringLinkNode)
Changed Length to count all nodes regardless of whether or not they have
a value; now accurately represents how many items are in the first chain
as it now includes the header.
Changed the length() function to count nodes starting with the one after
the header (the first with a value).
Removed GetNew as part of removing that function from IStringList.