New Class: StringSorter
Added a new set of classes called StringSorters, meant to handle
lexicographic comparisons between strings.
Added the JavaDefault and WindowsExplorer StringSorter subclasses, meant
to sort strings as Java and Windows Explorer do respectively.
Removed the defaultCaseSensitivity variable and replaced it with a
StringSorter object whose caseSensitive variable takes the place of said
variable.
Removed the old, incomplete, thread-unsafe logic around compareStrings,
replaced it with a call to the StringSorter object, and instituted a
revamped version of the logic into WindowsExplorer.
Expanded on the logic of getPrintable so that special whitespace can be
preserved and other "non-printable" characters can be escaped rather
than dropped outright. Removed toReadable, as getReadable now embodies
all of the logic it contained.
Revamped the logic of trimLeadingCharacters and trimTrailingCharacters,
moving away from going through "target" strings character-by-character
and instead grabbing the first/last character and seeing if it's in said
string.
Added isPrintable and isWritable functions that take an integer so that
chars can be tested rather than just complete strings.
Replaced two usages of indexOf,-1 with a call to contains.
Added JavaDoc and comments to various functions.