Made throwing on writing unreadables optional
Added variables to FileHandler that determine whether or not an
exception is thrown when a value outside of what is read from a file (0
to 255, called "writable" in the code) is written to file (the idea
being, only write what can be read back out).
Added functionality, along with a controlling variable, for escaping
strings into a printable state when they are written to file (thus
avoiding the writing of an "unwritable" character). Distinct from
escape/unescape, which focuses on storing and retrieving strings with
unprintable characters rather than making strings human-readable.