toAlphanumeric cleanup
Did some cleanup/enhancement of toAlphanumeric.
Fixed an issue in toAlphanumeric and commaDelineateNumber where the
leading "sign" character was not added properly to the result.
Removed the non-runtime exception from unescape so that functions that
make use of it don't have to explicitly declare the possibility of error
(leave it to be caught by high-level catch blocks).