Null handling with CompareStrings
Added handling to the base CompareStrings function to handle nulls.
If handle-nulls is specified, will treat nulls as legitimate strings: a
populated string is an occupied bucket, an empty string is an empty
bucket, and a null is a missing bucket; thus, as a missing bucket weighs
less than empty and occupied buckets, nulls come before empty and
populated strings.
If handle-nulls is turned down, makes sure the null is passed on to the
compareStringsAfterCase function.
Added an argument to the constructor and adjusted subclass constructors
to accommodate.